fix: rename wrong filename localStoradre -> to localStorage, and actualize paths of import
