+++ src/components/auth/forms/LoginForm.tsx
@@ -2,7 +2,7 @@ import 'react-app-polyfill/ie11'
 import * as React from 'react'
 import { useState } from 'react'
 import { useFormik } from 'formik'
-import { localStorageRemove } from '../../../utils/localStoradre'
+import { localStorageRemove } from '../../../utils/localStorage'
 import { authApi } from 'src/services/AuthService'
 import { useNavigate } from 'react-router-dom'
 import { useAppDispatch } from 'src/hooks/redux'
