+++ src/components/auth/forms/RegisterForm.tsx
@@ -7,7 +7,7 @@ import { useAppDispatch } from 'src/hooks/redux'
 import { validateRegister } from './formValidator'
 import { userAPI } from '../../../services/UserService'
 import { IUserForm } from '../../../models/IUser'
-import { localStorageRemove } from 'src/utils/localStoradre'
+import { localStorageRemove } from 'src/utils/localStorage'
 import { authApi } from '../../../services/AuthService'
 import styles from './Form.module.scss'
 import { authSlice } from '../../../store/reducers/authSlice'
