+++ src/components/games/GamesOverScreen.tsx
@@ -4,7 +4,7 @@ import styles from './GamesOverScreen.module.scss'
 import { useAppSelector } from '../../hooks/redux'
 import { useNavigate } from 'react-router-dom'
 import { userWordsAPI } from 'src/services/UsersWordsService'
-import { localStorageGet } from 'src/utils/localStoradre'
+import { localStorageGet } from 'src/utils/localStorage'
 import { getDateFromRu, getNowDateRu } from 'src/utils/date-helper'
 import { IDateStat, IUsersWords } from 'src/models/IUsersWords'
 import { IWord } from 'src/models/IWord'
