+++ src/components/textbookPage/TextbookPage.tsx
@@ -5,8 +5,8 @@ import { clsx } from 'src/utils/clsx'
 import BookSubpage from './bookSubpage/BookSubpage'
 import HardSubpage from './hardSubpage/HardSubpage'
 import CompleteSubpage from './completeSubpage/CompleteSubpage'
-import { Link } from 'react-router-dom'
-import { localStorageGet } from '../../utils/localStoradre'
+import { Link, useNavigate } from 'react-router-dom'
+import { localStorageGet } from '../../utils/localStorage'
 import { userWordsAPI } from '../../services/UsersWordsService'
 import { useAppSelector } from '../../hooks/redux'
 import { wordsAPI } from '../../services/WordsService'
