+++ src/components/textbookPage/detail/Detail.tsx
@@ -4,7 +4,7 @@ import { clsx } from 'src/utils/clsx'
 import styles from './Detail.module.scss'
 import { useAudio } from '../../../hooks/useAudio'
 import { userWordsAPI } from '../../../services/UsersWordsService'
-import { localStorageGet } from '../../../utils/localStoradre'
+import { localStorageGet } from '../../../utils/localStorage'
 import { useAppSelector } from '../../../hooks/redux'
 import { useDispatch } from 'react-redux'
 import { userWordsSlice } from '../../../store/reducers/userWordsSlice'
